feat: Create App composable
